requirements.txt
streamlit
pandas
scikit-learn
matplotlib
reportlab
google-generativeai
groq
